third_party/WebKit/LayoutTests/typedcssom/scaletransformcomponent.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
var values = [
  {input: new ScaleTransformComponent(0, 0), x: 0, y: 0, z: 1, is2DComponent: true},
  {input: new ScaleTransformComponent(1, 2), x: 1, y: 2, z: 1, is2DComponent: true},
  {input: new ScaleTransformComponent(-2, -4), x: -2, y: -4, z: 1, is2DComponent: true},
  {input: new ScaleTransformComponent(3.4, 2.7), x: 3.4, y: 2.7, z: 1, is2DComponent: true},
  {input: new ScaleTransformComponent(0, 0, 0), x: 0, y: 0, z: 0, is2DComponent: false},
  {input: new ScaleTransformComponent(1, 2, 3), x: 1, y: 2, z: 3, is2DComponent: false},
  {input: new ScaleTransformComponent(3.5, -2.7, -2), x: 3.5, y: -2.7, z: -2, is2DComponent: false}
];

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_equals(values[i].input.x, values[i].x);
    assert_equals(values[i].input.y, values[i].y);
    assert_equals(values[i].input.z, values[i].z);
  }
}, "Test that the (x, y, z) values for ScaleTransformComponent are correct.");

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_equals(values[i].input.is2DComponent(), values[i].is2DComponent);
  }
}, "Test that the is2DComponent values for ScaleTransformComponent is correct.");

</script>

<body>
</body>
